split out exceptions, command line run logic, and streams from libroll main file
