+++ libroll/__init__.py
@@ -1,1 +1,4 @@
+from .cmd import *
+from .streams import *
+from .exceptions import *
 from .libroll import *
